requirements.txt
pandas
gspread
google.oauth2.service_account
streamlit